openMV 主板带摄像头OV5640可以正常工作，RGB565格式可以显示到LCD上，JPEG格式通过USB cdc发送到上位机可以正常显示

+++ remake_openMV/SPI_LCD/lcd_display.c
@@ -11,7 +11,7 @@
 
 
 
-void lcd_display_image(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint8_t *image)
+void lcd_display_image(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint16_t *image)
 {
 	ST7735_DrawImage(x,y,width,height,image);
