+++ remake_openMV/SPI_LCD/lcd_display.h
@@ -14,7 +14,7 @@
 
 
 void lcd_init(void);
-void lcd_display_image(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint8_t *image);
+void lcd_display_image(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint16_t *image);
 void LCD_ShowString(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint8_t size, uint8_t *p);
 void LCD_ShowxNum(uint16_t x, uint16_t y, uint32_t num, uint8_t len, uint8_t size, uint8_t mode);
 void LCD_ShowNum(uint16_t x, uint16_t y, uint32_t num, uint8_t len, uint8_t size);
